jd-resume-analyzer/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
rapidfuzz>=3.6.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
rich>=13.7.0
PyPDF2>=3.0.0